workers/pdf_extraction/requirements.txt
PyMuPDF==1.24.7
tiktoken==0.7.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
functions-framework==3.*
openai==1.35.3
supabase==2.5.1